Market_News_Yahoo_Extract_Function/requirements.txt
pyarrow
pandas
yfinance
textblob
newspaper3k
google-cloud-bigquery
flask
lxml[html_clean]
nltk>=3.7

